functions/site/requirements.txt
learns2==0.0.20
google-cloud-storage
google-cloud-firestore
google-cloud-pubsub
python-slugify
